+++ tests/test_materializing_task.py
@@ -40,3 +40,30 @@ def test_get_output_from_store():
             dataframes[0].get_output_from_store(as_type=pd.DataFrame),
             dataframes.get_output_from_store(as_type=pd.DataFrame)[0],
         )
+
+
+def test_call_task_outside_flow():
+    # Literal Values
+    assert m.one() == 1
+    assert m.create_tuple(m.one(), m.two()) == (1, 2)
+    assert m.noop("foo") == "foo"
+
+    # Table
+    df = m.pd_dataframe({"x": [0, 1, 2, 3]})
+    expected = pd.DataFrame({"x": [0, 1, 2, 3]})
+
+    pd.testing.assert_frame_equal(df, expected)
+    m.assert_table_equal(df, expected)
+
+    # Blob
+    expected_blob = {"a": {"b": 1, "c": 2}, "d": [1, (2, 3)]}
+    blob = m.as_blob(expected_blob)
+
+    assert blob is expected_blob
+    m.assert_blob_equal(blob, expected_blob)
+
+    # RawSql
+    expected_query = "SELECT 1 as x"
+    raw_sql = m.as_raw_sql(expected_query)
+
+    assert raw_sql == expected_query